Disable proguard
For some reason it causes getAllDevices() to throw Exception. First it looks
like the server url is empty, but it seems the server url is there the first
time the method runs. An exception is probably thrown because of GSON missing
classes which proguard has deleted, and this causes a retry with the empty
server url. Or something.